json: switch library to jsoncpp, add JSON_SetDouble(), JSON_SetInteger(), JSON_SetString()

+++ src_cpp/tests/test_json.cpp
@@ -36,7 +36,6 @@ TEST_CASE( "json_open", "[json]" ) {
         REQUIRE(ptr != nullptr);
         REQUIRE(JSON_FileClose(ptr));
         REQUIRE(std::filesystem::exists(filename));
-        std::filesystem::remove(filename);
     }
 }
 
@@ -61,7 +60,6 @@ TEST_CASE( "json_set_bool", "[json]" ) {
 
         REQUIRE(JSON_FileClose(ptr));
         REQUIRE(std::filesystem::exists(filename));
-        std::filesystem::remove(filename);
     }
 
     SECTION("level_2") {
@@ -76,11 +74,91 @@ TEST_CASE( "json_set_bool", "[json]" ) {
 
         REQUIRE(JSON_SetBoolW(ptr, L"/meow/gaw", true));
         REQUIRE(JSON_SetBoolW(ptr, L"/meow/mur", true));
-        REQUIRE(JSON_SetBoolW(ptr, L"/gaw", true));
+        REQUIRE(JSON_SetBoolW(ptr, L"/kukareku", true));
 
         REQUIRE(JSON_FileClose(ptr));
         REQUIRE(std::filesystem::exists(filename));
-        std::filesystem::remove(filename);
     }
 }
 
+
+TEST_CASE( "json_set_double", "[json]" ) {
+    SECTION("nullptr") {
+        REQUIRE_FALSE(JSON_SetDoubleW(nullptr, L"meow", 1.2));
+        REQUIRE_FALSE(JSON_SetDoubleW(reinterpret_cast<void*>(0x1234), nullptr, 1.2));
+    }
+
+    SECTION("level_1") {
+        const wchar_t* filename = L"set_double_level_1.json";
+        if(std::filesystem::exists(filename)){
+            std::filesystem::remove(filename);
+        }
+        auto* ptr = JSON_FileOpenW(filename, true);
+        REQUIRE(ptr != nullptr);
+
+        REQUIRE_FALSE(JSON_SetDoubleW(ptr, L"gaw", 1.1));
+        REQUIRE_FALSE(JSON_SetDoubleW(ptr, L"gaw/", 1.2));
+        REQUIRE_FALSE(JSON_SetDoubleW(ptr, L"/gaw/", 1.3));
+
+        REQUIRE(JSON_SetDoubleW(ptr, L"/gaw", 1.3));
+        REQUIRE(JSON_SetDoubleW(ptr, L"/meow", 1.5));
+
+        REQUIRE(JSON_FileClose(ptr));
+        REQUIRE(std::filesystem::exists(filename));
+    }
+}
+
+
+TEST_CASE( "json_set_integer", "[json]" ) {
+    SECTION("nullptr") {
+        REQUIRE_FALSE(JSON_SetIntegerW(nullptr, L"meow", 1));
+        REQUIRE_FALSE(JSON_SetIntegerW(reinterpret_cast<void*>(0x1234), nullptr, 1));
+    }
+
+    SECTION("level_1") {
+        const wchar_t* filename = L"set_integer_level_1.json";
+        if(std::filesystem::exists(filename)){
+            std::filesystem::remove(filename);
+        }
+        auto* ptr = JSON_FileOpenW(filename, true);
+        REQUIRE(ptr != nullptr);
+
+        REQUIRE_FALSE(JSON_SetIntegerW(ptr, L"gaw", 1));
+        REQUIRE_FALSE(JSON_SetIntegerW(ptr, L"gaw/", 1));
+        REQUIRE_FALSE(JSON_SetIntegerW(ptr, L"/gaw/", 1));
+
+        REQUIRE(JSON_SetIntegerW(ptr, L"/gaw", 1));
+        REQUIRE(JSON_SetIntegerW(ptr, L"/meow", 2));
+
+        REQUIRE(JSON_FileClose(ptr));
+        REQUIRE(std::filesystem::exists(filename));
+    }
+}
+
+
+TEST_CASE( "json_set_string", "[json]" ) {
+    SECTION("nullptr") {
+        REQUIRE_FALSE(JSON_SetStringW(nullptr, L"meow", L"gaw"));
+        REQUIRE_FALSE(JSON_SetStringW(reinterpret_cast<void *>(0x1234), nullptr, L"gaw"));
+        REQUIRE_FALSE(JSON_SetStringW(reinterpret_cast<void *>(0x1234), L"meow", nullptr));
+    }
+
+    SECTION("level_3") {
+        const wchar_t* filename = L"set_string_level_3.json";
+        if(std::filesystem::exists(filename)){
+            std::filesystem::remove(filename);
+        }
+        auto* ptr = JSON_FileOpenW(filename, true);
+        REQUIRE(ptr != nullptr);
+
+        REQUIRE_FALSE(JSON_SetStringW(ptr, L"meow/gaw", L"one"));
+        REQUIRE_FALSE(JSON_SetStringW(ptr, L"/meow/gaw", nullptr));
+
+        REQUIRE(JSON_SetStringW(ptr, L"/meow/gaw", L"two"));
+        REQUIRE(JSON_SetStringW(ptr, L"/kukareku", L"two"));
+        REQUIRE(JSON_SetStringW(ptr, L"/kek/lol/azaza", L"three"));
+
+        REQUIRE(JSON_FileClose(ptr));
+        REQUIRE(std::filesystem::exists(filename));
+    }
+}
